[JC/Week6] TabBar 연결, HomeView를 앱의 시작으로 설정
