+++ jc/Week6_Mission/Week6_Mission/HomeView/TopInfoView.swift
@@ -1,0 +1,45 @@
+//
+//  WidthSizeRectangle.swift
+//  Week6_Mission
+//
+//  Created by Jiwoong CHOI on 5/14/24.
+//
+
+import SwiftUI
+
+struct TopInfoView: View {
+  var body: some View {
+    ZStack {
+      Rectangle()
+        .foregroundColor(.mint)
+        .frame(width: 400, height: 150, alignment: .center)
+        .clipShape(
+          UnevenRoundedRectangle(
+            cornerRadii: .init(topLeading: 0, bottomLeading: 20, bottomTrailing: 20, topTrailing: 0)
+          ))
+
+      VStack {
+        HStack {
+          Text("우리집")
+          Spacer()
+
+          Image(systemName: "square.grid.2x2")
+          Image(systemName: "bell")
+          Image(systemName: "cart")
+        }
+        .padding([.leading, .trailing], 15)
+        .padding(.top, 40)
+
+        Rectangle()
+          .fill(Color.white)
+          .frame(width: 370, height: 40)
+          .border(Color.black)
+      }
+    }
+    .ignoresSafeArea()
+  }
+}
+
+#Preview {
+  TopInfoView()
+}
