+++ jc/Week6_Mission/Week6_Mission/TabBarView/MyPageView.swift
@@ -1,0 +1,18 @@
+//
+//  MyPageView.swift
+//  Week6_Mission
+//
+//  Created by Jiwoong CHOI on 5/14/24.
+//
+
+import SwiftUI
+
+struct MyPageView: View {
+  var body: some View {
+    Text( "My Page View")
+  }
+}
+
+#Preview {
+  MyPageView()
+}
