+++ jc/Week6_Mission/Week6_Mission/TabBarView/MyView.swift
@@ -1,0 +1,18 @@
+//
+//  MyView.swift
+//  Week6_Mission
+//
+//  Created by Jiwoong CHOI on 5/14/24.
+//
+
+import SwiftUI
+
+struct MyView: View {
+  var body: some View {
+    Text( "My View")
+  }
+}
+
+#Preview {
+  MyView()
+}
