+++ jc/Week6_Mission/Week6_Mission/TabBarView/OrderView.swift
@@ -1,0 +1,18 @@
+//
+//  OrderView.swift
+//  Week6_Mission
+//
+//  Created by Jiwoong CHOI on 5/14/24.
+//
+
+import SwiftUI
+
+struct OrderView: View {
+  var body: some View {
+    Text( "Order View")
+  }
+}
+
+#Preview {
+  OrderView()
+}
