+++ jc/Week6_Mission/Week6_Mission/TabBarView/SearchView.swift
@@ -1,0 +1,18 @@
+//
+//  SearchView.swift
+//  Week6_Mission
+//
+//  Created by Jiwoong CHOI on 5/14/24.
+//
+
+import SwiftUI
+
+struct SearchView: View {
+  var body: some View {
+    Text( "Search View")
+  }
+}
+
+#Preview {
+  SearchView()
+}
